+++ unifi_console/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG - Unifi Console
 
+
+## 1.2.0
+* [Added] Add support for device specific metrics (UAP, UDM, USG, USW, USX). See [#1715](https://github.com/DataDog/integrations-extras/pull/1715)
+* [Fix] Fix checks pilling up
+
 ## 1.1.1
 * [Fix] Fix url issue for non UMP setup. See [#1596](https://github.com/DataDog/integrations-extras/pull/1596)